requirements/base.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.3
psycopg==3.1.12
channels==4.0.0
channels-redis==4.1.0
Pillow==10.0.1
celery==5.3.4
redis==5.0.1
django-treebeard==4.7